ofmcc-3826 - make buildDateFilter function to be more generic
